requirements.txt
discord.py
asyncio==3.4.3
python-dateutil==2.8.0